{"schema":"maestro/commit/1","summary":"extract overlay frozen text and ime runtime module","authority":"XY-244"}
